refactor: [code-review] Delete the dead checkoutless `HubCoordinationEx… [ORB-11599]
[code-review] Delete the dead checkoutless `HubCoordinationExecutor` task/friction CRUD path

Planned-By: grok
Orchestrated-By: astra
