+++ crates/orbit-core/src/adapter/tool_host/tests/task_tools.rs
@@ -1,15 +1,12 @@
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Barrier, Mutex, MutexGuard, OnceLock};
+use std::sync::{Mutex, MutexGuard, OnceLock};
 
-use orbit_store::maintenance::task_registry::{
-    TaskRegistryStore, read_workspace_config, task_registry_path,
-};
+use orbit_store::maintenance::task_registry::read_workspace_config;
 use orbit_types::task::{TASK_SHOW_PUBLIC_DTO_FIELDS, TaskStatus};
 use orbit_types::tool::ToolSessionContext;
 use serde_json::{Value, json};
 
-use super::super::HubCoordinationExecutor;
 use super::super::json::task_to_json;
 use super::super::test_support::{
     create_task, create_task_with_crew, invalid_input_message, run_tool_as_operator, test_runtime,
@@ -64,72 +61,6 @@ fn assert_task_titles(output: &Value, expected: &[&str]) {
     expected.sort();
 
     assert_eq!(titles, expected);
-}
-
-fn checkoutless_executor() -> (
-    tempfile::TempDir,
-    HubCoordinationExecutor,
-    ToolSessionContext,
-) {
-    let root = tempfile::tempdir().expect("global root");
-    HubCoordinationExecutor::register_workspace(root.path(), "ws_checkoutless", "checkoutless")
-        .expect("register workspace");
-    let executor = HubCoordinationExecutor::new(root.path(), "ws_checkoutless", None)
-        .expect("coordination executor");
-    let context = ToolSessionContext::trusted_local(
-        Some("ws_checkoutless".to_string()),
-        Some("hm_hub".to_string()),
-        Some("hub".to_string()),
-    );
-    (root, executor, context)
-}
-
-fn checkoutless_review_task(
-    executor: &HubCoordinationExecutor,
-    context: &ToolSessionContext,
-    title: &str,
-) -> String {
-    let created = executor
-        .execute_tool(
-            "orbit.task.add",
-            json!({
-                "workspace": "ws_checkoutless",
-                "title": title,
-                "description": "Checkoutless review fixture",
-                "complexity": "low",
-                "model": "codex"
-            }),
-            context.clone(),
-        )
-        .expect("add checkoutless task");
-    let id = created["id"].as_str().expect("task id").to_string();
-    executor
-        .execute_tool(
-            "orbit.task.update",
-            json!({"id": id, "status": "backlog", "plan": "Execute", "model": "codex"}),
-            context.clone(),
-        )
-        .expect("move task to backlog");
-    executor
-        .execute_tool(
-            "orbit.task.start",
-            json!({"id": id, "model": "codex"}),
-            context.clone(),
-        )
-        .expect("start task");
-    executor
-        .execute_tool(
-            "orbit.task.update",
-            json!({
-                "id": id,
-                "status": "review",
-                "execution_summary": "Outcome: success",
-                "model": "codex"
-            }),
-            context.clone(),
-        )
-        .expect("move task to review");
-    id
 }
 
 #[test]
@@ -622,193 +553,6 @@ fn task_tools_roundtrip_required_tools_and_reject_updates() {
         )
         .expect_err("task requirements are creation-only");
     assert!(error.to_string().contains("immutable"), "{error}");
-}
-
-#[test]
-fn checkoutless_metadata_update_serializes_without_restoring_an_older_status() {
-    let (_root, executor, context) = checkoutless_executor();
-    let id = checkoutless_review_task(&executor, &context, "Terminal race fixture");
-    let snapshot_read = Arc::new(Barrier::new(2));
-    let release_update = Arc::new(Barrier::new(2));
-    executor.set_after_pre_state_read_hook(Arc::new({
-        let id = id.clone();
-        let snapshot_read = Arc::clone(&snapshot_read);
-        let release_update = Arc::clone(&release_update);
-        move |task, input| {
-            if task.id == id && input["title"] == "Stale title" {
-                assert_eq!(task.status, TaskStatus::Review);
-                snapshot_read.wait();
-                release_update.wait();
-            }
-        }
-    }));
-
-    std::thread::scope(|scope| {
-        let stale_executor = executor.clone();
-        let stale_context = context.clone();
-        let stale_id = id.clone();
-        let stale_update = scope.spawn(move || {
-            stale_executor.execute_tool(
-                "orbit.task.update",
-                json!({"id": stale_id, "title": "Stale title", "model": "codex"}),
-                stale_context,
-            )
-        });
-
-        snapshot_read.wait();
-        let approved = executor
-            .execute_tool(
-                "orbit.task.approve",
-                json!({"id": id, "note": "Approved normally", "model": "codex"}),
-                context.clone(),
-            )
-            .expect("approve review task through the normal transition surface");
-        assert_eq!(approved["status"], "done");
-        assert_eq!(
-            approved["comments"]
-                .as_array()
-                .and_then(|comments| comments.last())
-                .and_then(|comment| comment["message"].as_str()),
-            Some("Approved normally")
-        );
-        release_update.wait();
-
-        let updated = stale_update
-            .join()
-            .expect("stale update thread")
-            .expect("explicit metadata update remains valid after completion");
-        assert_eq!(updated["status"], "done");
-        assert_eq!(updated["title"], "Stale title");
-    });
-
-    let shown = executor
-        .execute_tool(
-            "orbit.task.show",
-            json!({"id": id, "fields": ["status", "title"]}),
-            context,
-        )
-        .expect("show terminal task");
-    assert_eq!(shown, json!({"status": "done", "title": "Stale title"}));
-}
-
-#[test]
-fn checkoutless_task_update_supports_direct_archive_and_restore_with_fields() {
-    let (_root, executor, context) = checkoutless_executor();
-    let id = checkoutless_review_task(&executor, &context, "Flexible status fixture");
-    let done = executor
-        .execute_tool(
-            "orbit.task.update",
-            json!({"id": id, "status": "done", "model": "codex"}),
-            context.clone(),
-        )
-        .expect("review moves directly to done");
-    assert_eq!(done["status"], "done");
-
-    let archived = executor
-        .execute_tool(
-            "orbit.task.update",
-            json!({
-                "id": id,
-                "status": "archived",
-                "title": "Archived classification",
-                "model": "codex"
-            }),
-            context.clone(),
-        )
-        .expect("done moves directly to archived with a field edit");
-    assert_eq!(archived["status"], "archived");
-    assert_eq!(archived["title"], "Archived classification");
-
-    let restored = executor
-        .execute_tool(
-            "orbit.task.update",
-            json!({"id": id, "status": "rejected", "model": "codex"}),
-            context,
-        )
-        .expect("archived restores directly to rejected");
-    assert_eq!(restored["status"], "rejected");
-    let history = restored["history"].as_array().expect("history array");
-    assert!(
-        history
-            .iter()
-            .any(|entry| { entry["from_status"] == "done" && entry["to_status"] == "archived" })
-    );
-    assert!(
-        history.iter().any(|entry| {
-            entry["from_status"] == "archived" && entry["to_status"] == "rejected"
-        })
-    );
-}
-
-#[test]
-fn checkoutless_updates_to_distinct_tasks_remain_independent() {
-    let (root, executor, context) = checkoutless_executor();
-    let first = checkoutless_review_task(&executor, &context, "First concurrent task");
-    let second = checkoutless_review_task(&executor, &context, "Second concurrent task");
-    let registry =
-        TaskRegistryStore::open(&task_registry_path(root.path())).expect("open task registry");
-    let first_lock_path = registry
-        .canonical_task_bundle_path("ws_checkoutless", &first)
-        .expect("first task bundle")
-        .join(".task.yaml.lock");
-    let first_locked = Arc::new(Barrier::new(2));
-    let release_first = Arc::new(Barrier::new(2));
-    executor.set_after_locked_state_read_hook(Arc::new({
-        let first = first.clone();
-        let first_locked = Arc::clone(&first_locked);
-        let release_first = Arc::clone(&release_first);
-        move |task, input| {
-            if task.id == first && input["title"] == "First task updated" {
-                first_locked.wait();
-                release_first.wait();
-            }
-        }
-    }));
-
-    std::thread::scope(|scope| {
-        let first_executor = executor.clone();
-        let first_context = context.clone();
-        let first_id = first.clone();
-        let first_update = scope.spawn(move || {
-            first_executor.execute_tool(
-                "orbit.task.update",
-                json!({
-                    "id": first_id,
-                    "title": "First task updated",
-                    "model": "codex"
-                }),
-                first_context,
-            )
-        });
-
-        first_locked.wait();
-        let first_lock = std::fs::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .open(&first_lock_path)
-            .expect("open first task lock file");
-        let error = fs2::FileExt::try_lock_exclusive(&first_lock)
-            .expect_err("the updater must hold the task lock after its authoritative read");
-        assert_eq!(error.kind(), std::io::ErrorKind::WouldBlock);
-        let second_updated = executor
-            .execute_tool(
-                "orbit.task.update",
-                json!({
-                    "id": second,
-                    "title": "Second task updated",
-                    "model": "codex"
-                }),
-                context.clone(),
-            )
-            .expect("a distinct task can update while the first task lock is held");
-        assert_eq!(second_updated["title"], "Second task updated");
-        release_first.wait();
-        let first_updated = first_update
-            .join()
-            .expect("first update thread")
-            .expect("first update succeeds after release");
-        assert_eq!(first_updated["title"], "First task updated");
-    });
 }
 
 /// ORB-10968: crew configuration is host-local, so a stored crew this host has
